crewai_service/pyproject.toml
[project]
name = "first-principles-crew"
version = "0.1.0"
description = "First Principles Thinking Coach CrewAI Service"
authors = [
    {name = "First Principles Team", email = "team@firstprinciples.ai"}
]
dependencies = [
    "crewai>=0.28.0",
    "crewai-tools>=0.5.0",
    "openai>=1.13.3",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "litellm>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
python = "3.12"

[tool.black]
line-length = 88
target-version = ['py312']

[tool.ruff]
line-length = 88
target-version = "py312"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]